Extend PureComponent in WebPreview
Previously, during the refactor I made in #6881, I imported the
`shallowCompare()` helper function to replace the `PureRenderMixin` that
was originally there.

In this PR, I am foregoing this import in favor of simply extending the
`PureComponent`, which further removes noise from inside this
already-big component.

Props to @aduthie for pointing me to this base class.